CI-89: refactoring styles
