+++ app/assets/stylesheets/main.scss
@@ -50,21 +50,6 @@ body {
   border-radius: 15px;
   padding-left: 23px;
 }
-.btn-next, .btn-next:disabled {
-  width: 235px;
-  height: 52px;
-  background: $light-blue;
-  border: 1px solid #000000;
-  box-shadow: 0 4px 4px rgba(0, 0, 0, 0.25);
-  border-radius: 12px;
-  font-family: 'Inter', sans-serif;
-  font-style: normal;
-  font-weight: 700;
-  font-size: 26.6341px;
-  line-height: 32px;
-  color: $color-white;
-  margin-left: 70px;
-}
 .fa-question {
   color:  $color-rose;
 }
