+++ app/javascript/components/UI/BackButton.js
@@ -1,5 +1,5 @@
 import React from 'react';
-import {backHandling} from "../helpers/helper";
+import {backHandling} from "../helpers/helpers";
 import {Button} from "react-bootstrap";
 
 const BackButton = ({data, setData}) => <Button onClick={backHandling}>Back</Button>
