--- app/javascript/components/helpers/helper.js
@@ -1,5 +1,0 @@
-export const isEmpty = (obj) => Object.keys(obj).length === 0 && obj.constructor === Object
-
-export const backHandling = () => {
-  window.history.back()
-}
