+++ app/views/passwordless/sessions/new.html.erb
@@ -19,7 +19,7 @@
     <div class="d-flex align-items-center" style="margin-top: 100px">
       <div class="w-100">
         <div class="justify-content-center">
-          <%= f.submit 'Next', class: "btn btn-next", id: "next-button", disabled: true %>
+          <%= f.submit 'Next', class: "primary-btn", style: 'margin-left: 70px', id: "next-button", disabled: true %>
         </div>
       </div>
       <div class="flex-shrink-1">
